Move parameters utility class to grass plugin
This is only used by the grass processing plugin
