+++ python/plugins/grassprovider/tests/CMakeLists.txt
@@ -10,4 +10,5 @@ if(ENABLE_TESTS)
   ADD_PYTHON_TEST(ProcessingGrassAlgorithmsRasterTestPt3 grass_algorithms_raster_test_pt3.py)
   ADD_PYTHON_TEST(ProcessingGrassAlgorithmsVectorTest grass_algorithms_vector_test.py)
   ADD_PYTHON_TEST(ProcessingGrassAlgsPostgreRasterProvider grass_algorithm_postgreraster_test.py)
+  ADD_PYTHON_TEST(ProcessingParametersTest ParametersTest.py)
 endif()
