+++ python/plugins/processing/tests/CMakeLists.txt
@@ -9,7 +9,6 @@ if(ENABLE_TESTS)
   ADD_PYTHON_TEST(ProcessingGeneralTest ProcessingGeneralTest.py)
   ADD_PYTHON_TEST(ProcessingGuiTest GuiTest.py)
   ADD_PYTHON_TEST(ProcessingProjectProviderTest ProjectProvider.py)
-  ADD_PYTHON_TEST(ProcessingParametersTest ParametersTest.py)
   ADD_PYTHON_TEST(ProcessingGenericAlgorithmsTest AlgorithmsTestBase.py)
   ADD_PYTHON_TEST(ProcessingQgisAlgorithmsTestPt1 QgisAlgorithmsTest1.py)
   ADD_PYTHON_TEST(ProcessingQgisAlgorithmsTestPt2 QgisAlgorithmsTest2.py)
